AdEase Time Series
Ad Ease is an ads and marketing-based company helping businesses elicit maximum clicks @ minimum cost. AdEase is an ad infrastructure to help businesses promote themselves easily, effectively, and economically. The interplay of 3 AI modules - Design, Dispense, and Decipher, come together to make it an end-to-end 3-step process digital advertising solution for all.

You are working in the Data Science team of Ad Ease trying to understand the per-page view report for different Wikipedia pages for 550 days, and forecasting the number of views so that you can predict and optimize the ad placement for your clients. You are provided with the data of 145k Wikipedia pages and the daily view count for each of them. Your clients belong to different regions and need data on how their ads will perform on pages in different languages.